Oh... I see...
There can be conflicts between hotkeys that are only in the normal
game and hotkeys that are only in the campaign (I think that is what
is going on).  Perhaps while editing the hotkeys in the campaign
I create unseen conflicts in the normal game.

I did discover that you must completely exit the game before you can
edit the normal game hotkeys once you have entered the campaign.